Undid delete of commented out tests and moved new tests
