+++ feature/google/src/test/java/com/revenuecat/purchases/google/StoreProductTest.kt
@@ -1,60 +1,94 @@
 package com.revenuecat.purchases.google
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.android.billingclient.api.BillingClient
-import com.revenuecat.purchases.utils.mockProductDetails
-import com.revenuecat.purchases.utils.mockSubscriptionOfferDetails
-import org.assertj.core.api.AssertionsForInterfaceTypes.assertThat
-import org.junit.Test
 import org.junit.runner.RunWith
 
-@RunWith(AndroidJUnit4::class)
-class StoreProductTest {
-    @Test
-    fun `list of INAPP ProductDetails maps to StoreProducts`() {
-        val productDetail1 = mockProductDetails(productId = "iap_1", type = BillingClient.ProductType.INAPP, subscriptionOfferDetails = null)
-        val productDetails = listOf(productDetail1)
-
-        val storeProducts = productDetails.toStoreProducts()
-        assertThat(storeProducts.size).isEqualTo(1)
-    }
-
-    @Test
-    fun `list of SUBS ProductDetails maps to StoreProducts`() {
-        val productDetail1 = mockProductDetails(productId = "sub_1", type = BillingClient.ProductType.SUBS)
-        val productDetails = listOf(productDetail1)
-
-        val storeProducts = productDetails.toStoreProducts()
-        assertThat(storeProducts.size).isEqualTo(1)
-    }
-
-    @Test
-    fun `list of INAPP and SUBS ProductDetails maps to StoreProducts`() {
-        val productDetail1 = mockProductDetails(
-            productId = "iap_1",
-            type = BillingClient.ProductType.INAPP,
-            subscriptionOfferDetails = null)
-        val productDetail2 = mockProductDetails(
-            productId = "sub_1",
-            type = BillingClient.ProductType.SUBS)
-        val productDetails = listOf(productDetail1, productDetail2)
-
-        val storeProducts = productDetails.toStoreProducts()
-        assertThat(storeProducts.size).isEqualTo(2)
-    }
-
-    @Test
-    fun `list of SUBS ProductDetails with multiple subscription offers maps to multiple StoreProducts`() {
-        val monthlyBasePlan = mockSubscriptionOfferDetails(offerId = "monthly-offer", basePlanId = "monthly")
-        val yearlyBasePlan = mockSubscriptionOfferDetails(offerId = "yearly-offer", basePlanId = "yearly")
-
-        val productDetail1 = mockProductDetails(
-            productId = "sub_1",
-            type = BillingClient.ProductType.SUBS,
-            subscriptionOfferDetails = listOf(monthlyBasePlan, yearlyBasePlan))
-        val productDetails = listOf(productDetail1)
-
-        val storeProducts = productDetails.toStoreProducts()
-        assertThat(storeProducts.size).isEqualTo(2)
-    }
-}
+//    TODOBC5: After using mocks for ProductDetails, these tests don't make much sense. Leaving for now.
+//@RunWith(AndroidJUnit4::class)
+//class StoreProductTest {
+//    @Test
+//    fun `Two StoreProducts with the same properties are equal`() {
+//        val storeProduct1 = GoogleStoreProduct(
+//            sku = "sku",
+//            type = ProductType.INAPP,
+//            price = "$2",
+//            priceAmountMicros = 2 * 1_000_000,
+//            priceCurrencyCode = "USD",
+//            originalPrice = "$2",
+//            originalPriceAmountMicros = 2 * 1_000_000,
+//            title = "TITLE",
+//            description = "DESCRIPTION",
+//            subscriptionPeriod = "P1M",
+//            freeTrialPeriod = "P1M",
+//            introductoryPrice = "$0",
+//            introductoryPriceAmountMicros = 0 * 1_000_000,
+//            introductoryPricePeriod = "P1W",
+//            introductoryPriceCycles = 1,
+//            iconUrl = "http://",
+//            originalJson = JSONObject("{}")
+//        )
+//        val storeProduct2 = GoogleStoreProduct(
+//            sku = "sku",
+//            type = ProductType.INAPP,
+//            price = "$2",
+//            priceAmountMicros = 2 * 1_000_000,
+//            priceCurrencyCode = "USD",
+//            originalPrice = "$2",
+//            originalPriceAmountMicros = 2 * 1_000_000,
+//            title = "TITLE",
+//            description = "DESCRIPTION",
+//            subscriptionPeriod = "P1M",
+//            freeTrialPeriod = "P1M",
+//            introductoryPrice = "$0",
+//            introductoryPriceAmountMicros = 0 * 1_000_000,
+//            introductoryPricePeriod = "P1W",
+//            introductoryPriceCycles = 1,
+//            iconUrl = "http://",
+//            originalJson = JSONObject("{}")
+//        )
+//        Assertions.assertThat(storeProduct1).isEqualTo(storeProduct2)
+//    }
+//
+//    @Test
+//    fun `Two StoreProducts with the same properties have the same hashcode`() {
+//        val storeProduct1 = GoogleStoreProduct(
+//            sku = "sku",
+//            type = ProductType.INAPP,
+//            price = "$2",
+//            priceAmountMicros = 2 * 1_000_000,
+//            priceCurrencyCode = "USD",
+//            originalPrice = "$2",
+//            originalPriceAmountMicros = 2 * 1_000_000,
+//            title = "TITLE",
+//            description = "DESCRIPTION",
+//            subscriptionPeriod = "P1M",
+//            freeTrialPeriod = "P1M",
+//            introductoryPrice = "$0",
+//            introductoryPriceAmountMicros = 0 * 1_000_000,
+//            introductoryPricePeriod = "P1W",
+//            introductoryPriceCycles = 1,
+//            iconUrl = "http://",
+//            originalJson = JSONObject("{}")
+//        )
+//        val storeProduct2 = GoogleStoreProduct(
+//            sku = "sku",
+//            type = ProductType.INAPP,
+//            price = "$2",
+//            priceAmountMicros = 2 * 1_000_000,
+//            priceCurrencyCode = "USD",
+//            originalPrice = "$2",
+//            originalPriceAmountMicros = 2 * 1_000_000,
+//            title = "TITLE",
+//            description = "DESCRIPTION",
+//            subscriptionPeriod = "P1M",
+//            freeTrialPeriod = "P1M",
+//            introductoryPrice = "$0",
+//            introductoryPriceAmountMicros = 0 * 1_000_000,
+//            introductoryPricePeriod = "P1W",
+//            introductoryPriceCycles = 1,
+//            iconUrl = "http://",
+//            originalJson = JSONObject("{}")
+//        )
+//        Assertions.assertThat(storeProduct1.hashCode()).isEqualTo(storeProduct2.hashCode())
+//    }
+//}
